requirements.txt
django==3.2
psycopg2-binary
djangorestframework
requests